CLINICAL-SUPPLY/requirements.txt
# Core dependencies
python-dotenv>=1.0.0
pandas>=2.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
streamlit>=1.28.0
requests>=2.31.0
chardet>=5.0.0

# Gemini API
google-generativeai>=0.3.0

# AgentOps
agentops

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# MCP & A2A Integration
websockets>=12.0

